fix: memory leak

+++ module-07/ex02/Array.hpp
@@ -7,8 +7,8 @@ template <typename T>
 class Array
 {
 	private:
-		T*		_array;
-		size_t	_size;
+		T*				_array;
+		unsigned int	_size;
 
 	public:
 		Array();
@@ -18,10 +18,9 @@ class Array
 		Array&	operator=(const Array& other);
 		T&		operator[](unsigned int index) const;
 		T* 		getArray() const;
-		size_t	size() const;
+		unsigned int	size() const;
 };
 
 #include "Array.tpp"
 
 #endif
-
